% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts.R
\name{impacts}
\alias{impacts}
\alias{impacts.SEjags}
\title{Compute impacts from a Bayesian spatial econometrics model ftited with SEjags.}
\usage{
impacts(obj, ...)

\method{impacts}{SEjags}(res, W)
}
\arguments{
\item{obj}{A SEjags object.}

\item{W}{An adjacency matrix, same as used in the call to SEjags().}
}
\value{
A named list with MCMC objects for direct, indirect and 
total impacts.
}
\description{
This function will compute the impacts (direct, indirect and total) from
a SEjags object.
}
\examples{
data(columbus)
W <- nb2mat(col.gal.nb, style = "W")
m.form <-  CRIME ~ INC + HOVAL
slm.mcmc <- SEjags(m.form, data = d, W = W, model = "slm")
impacts(slm.mcmc, W)
data(columbus)
W <- nb2mat(col.gal.nb, style = "W")
m.form <-  CRIME ~ INC + HOVAL
slm.mcmc <- SEjags(m.form, data = d, W = W, model = "slm")
impacts(slm.mcmc, W)
}
\keyword{models}
\keyword{spatial}

