% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts.R
\name{impacts.SEMCMC.sem}
\alias{impacts.SEMCMC.sem}
\alias{impacts.SEMCMC.slm}
\alias{impacts.SEMCMC.sdm}
\alias{impacts.SEMCMC.sdem}
\title{Compute impacts for different models}
\usage{
\method{impacts}{SEMCMC.sem}(obj, idx.var, var.names)

\method{impacts}{SEMCMC.slm}(obj, W, idx.var, var.names)

\method{impacts}{SEMCMC.sdm}(obj, W, idx.var, var.names)

\method{impacts}{SEMCMC.sdem}(obj, W, idx.var, var.names)
}
\arguments{
\item{obj}{A SEMCMC object.}

\item{idx.var}{An index to subset the covariates coeffiecients}

\item{var.names}{Vector with variable names}

\item{W}{Adjacency matrix.}
}
\description{
This is an internal function to compute the impacts (direct, indirect 
and total)  from a SEMCMC object.
}
\examples{
data(columbus)

W <- nb2mat(col.gal.nb, style = "W")
m.form <-  CRIME ~ INC + HOVAL
#SEM model
sem.mcmc <- SEMCMC(m.form, data = columbus, W = W, model = "sem")
impacts(sem.mcmc, W)
#SLM model
slm.mcmc <- SEMCMC(m.form, data = columbus, W = W, model = "slm")
impacts(slm.mcmc, W)
sdm.mcmc <- SEMCMC(m.form, data = columbus, W = W, model = "sdm")
impacts(sdm.mcmc, W)
data(columbus)
#SDEM model
sdem.mcmc <- SEMCMC(m.form, data = columbus, W = W, model = "sdem")
impacts(sdem.mcmc, W)
}
\keyword{models}
\keyword{spatial}
