% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.impacts.SEjags}
\alias{print.impacts.SEjags}
\title{Display summary of impacts from fitted with SEjags.}
\usage{
\method{print}{impacts.SEjags}(obj)
}
\arguments{
\item{obj}{A SEjags.impacts object.}
}
\description{
This function will print  summary of the impacts (direct, indirect and total) from
a SEjags object. Output is similar to spdep::impacts
}
\examples{
data(columbus)
W <- nb2mat(col.gal.nb, style = "W")
m.form <-  CRIME ~ INC + HOVAL
slm.mcmc <- SEjags(m.form, data = d, W = W, model = "slm")
impacts(slm.mcmc, W)
}
\keyword{models}
\keyword{spatial}

