% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEjags.R
\name{SEjags}
\alias{SEjags}
\title{Function to fit spatial econometrics models using MCMC with jags}
\usage{
SEjags(formula, data, W, model = "sem")
}
\arguments{
\item{formula}{Formula with response and covariates.}

\item{data}{Data.frame with the dataset.}

\item{W}{An adjacency matrix, same as used in the call to SEjags().}

\item{model.}{Model to be fitted: 'sem', 'slm', 'sdm', 'sdem', 'slx' or 
'sac'.}
}
\value{
A named list with MCMC objects as returned by jags.
}
\description{
This function will fit several spatial econometrics models
with jags. Models included are SEM, SLM, SDM, SDEM, SLX and SAC.
}
\examples{
data(columbus)
d <- columbus
W <- nb2mat(col.gal.nb, style = "W")
m.form <-  CRIME ~ INC + HOVAL

#Fit models
sem.mcmc <- SEjags(m.form, data = d, W = W, model = "sem")
slm.mcmc <- SEjags(m.form, data = d, W = W, model = "slm")
sdm.mcmc <- SEjags(m.form, data = d, W = W, model = "sdm")
sdem.mcmc <- SEjags(m.form, data = d, W = W, model = "sdem")
slx.mcmc <- SEjags(m.form, data = d, W = W, model = "slx")
sac.mcmc <- SEjags(m.form, data = d, W = W, model = "sac")

#Compute impacts
impacts(sem.mcmc, W)
impacts(slm.mcmc, W)
impacts(sdm.mcmc, W)
impacts(sdem.mcmc, W)
impacts(slx.mcmc, W)
impacts(sac.mcmc, W)
}
\keyword{models}
\keyword{spatial}

