% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts.R
\name{impacts}
\alias{impacts}
\alias{impacts.SEMCMC}
\title{Compute impacts from a Bayesian spatial econometrics model ftited with SEMCMC.}
\usage{
impacts(obj, ...)

\method{impacts}{SEMCMC}(obj, ...)
}
\arguments{
\item{obj}{A SEMCMC object.}

\item{...}{Extra argument to compute the impacts.}
}
\value{
A named list with MCMC objects for direct, indirect and 
total impacts.
}
\description{
This function will compute the impacts (direct, indirect and total) from
a SEMCMC object.
}
\examples{
data(columbus, package = "spdep")

W <- spdep::nb2mat(col.gal.nb, style = "W")
m.form <-  CRIME ~ INC + HOVAL
slm.mcmc <- SEMCMC(m.form, data = columbus, W = W, model = "slm")
impacts(slm.mcmc, W)
}
\keyword{models}
\keyword{spatial}
