% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impacts.R
\name{impacts.SEjags.sem}
\alias{impacts.SEjags.sdem}
\alias{impacts.SEjags.sdm}
\alias{impacts.SEjags.sem}
\alias{impacts.SEjags.slm}
\title{Compute impacts for different models}
\usage{
\method{impacts}{SEjags.sem}(res, idx.var, var.names)

\method{impacts}{SEjags.slm}(res, W, idx.var, var.names)

\method{impacts}{SEjags.sdm}(res, W, idx.var, var.names)

\method{impacts}{SEjags.sdem}(res, W, idx.var, var.names)
}
\arguments{
\item{res}{A SEjags object.}

\item{idx.var}{An index to subset the covariates coeffiecients}

\item{var.names}{Vector with variable names}

\item{W}{Adjacency matrix.}
}
\description{
This is an internal function to compute the impacts (direct, indirect 
and total)  from a SEjags object.
}
\examples{
data(columbus)
W <- nb2mat(col.gal.nb, style = "W")
m.form <-  CRIME ~ INC + HOVAL
#SEM model
sem.mcmc <- SEjags(m.form, data = d, W = W, model = "sem")
impacts(sem.mcmc, W)
#SLM model
slm.mcmc <- SEjags(m.form, data = d, W = W, model = "slm")
impacts(slm.mcmc, W)
sdm.mcmc <- SEjags(m.form, data = d, W = W, model = "sdm")
impacts(sdm.mcmc, W)
data(columbus)
#SDEM model
sdem.mcmc <- SEjags(m.form, data = d, W = W, model = "sdem")
impacts(sdem.mcmc, W)
}
\keyword{models}
\keyword{spatial}

