% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEjags.R
\name{SEjags}
\alias{SEjags}
\title{Function to fit spatial econometrics models using MCMC with jags}
\usage{
SEjags(formula, data, W, model = "sem", link = "identity",
  n.burnin = 1000, n.iter = 1000, n.thin = 1, linear.predictor = FALSE)
}
\arguments{
\item{formula}{Formula with response and covariates.}

\item{data}{Data.frame with the dataset.}

\item{W}{An adjacency matrix, same as used in the call to SEjags().}

\item{model}{Model to be fitted: 'sem', 'slm', 'sdm', 'sdem', 'slx',
'sac', 'sacmixed' (SAC with lagged covariates) or 'car'.}

\item{link}{One of 'indentity', 'logit' or 'probit'.}

\item{n.iter}{Number of iterarions after bun-in}

\item{n.thin}{Thinning interval}

\item{linear.predictor}{Whether the linear predictor should be saved (default
is FALSE).}

\item{n.burin}{Number of burn-in iterations}
}
\value{
A named list with MCMC objects as returned by jags.
}
\description{
This function will fit several spatial econometrics models
with jags. Models included are SEM, SLM, SDM, SDEM, SLX and SAC.
}
\examples{
data(columbus)
d <- columbus
W <- nb2mat(col.gal.nb, style = "W")
m.form <-  CRIME ~ INC + HOVAL

#Fit models with SEjags
sem.mcmc <- SEjags(m.form, data = d, W = W, model = "sem")
slm.mcmc <- SEjags(m.form, data = d, W = W, model = "slm")
sdm.mcmc <- SEjags(m.form, data = d, W = W, model = "sdm")
sdem.mcmc <- SEjags(m.form, data = d, W = W, model = "sdem")
slx.mcmc <- SEjags(m.form, data = d, W = W, model = "slx")
sac.mcmc <- SEjags(m.form, data = d, W = W, model = "sac")
sacmixed.mcmc <- SEjags(m.form, data = d, W = W, model = "sacmixed")
car.mcmc <- SEjags(m.form, data = d, W = W, model = "car")

#Compute impacts
impacts(sem.mcmc, W)
impacts(slm.mcmc, W)
impacts(sdm.mcmc, W)
impacts(sdem.mcmc, W)
impacts(slx.mcmc, W)
impacts(sac.mcmc, W)
impacts(sacmixed.mcmc, W)
impacts(car.mcmc, W)
}
\seealso{
\code{\link{lagsarlm}}, \code{\link{errorsarlm}} and
\code{\link{sacsarlm}} to fit similar models using maximum likelihood.
}
\keyword{models}
\keyword{spatial}

